Update: 后端 - 接口修改

+++ read-joy-api/src/main/java/com/readjoy/readjoyapi/service/impl/BookServiceImpl.java
@@ -10,19 +10,25 @@ import com.readjoy.readjoyapi.common.dto.book.SelectBookDTO;
 import com.readjoy.readjoyapi.common.dto.book.UpdateBookDTO;
 import com.readjoy.readjoyapi.common.mapper.BookMapper;
 import com.readjoy.readjoyapi.common.pojo.Book;
+import com.readjoy.readjoyapi.common.pojo.Resource;
 import com.readjoy.readjoyapi.common.utils.AssertUtil;
 import com.readjoy.readjoyapi.common.utils.LocalFileUtil;
 import com.readjoy.readjoyapi.common.vo.book.BookDetailVO;
 import com.readjoy.readjoyapi.common.vo.book.BookVO;
+import com.readjoy.readjoyapi.common.vo.resource.AdminResourceVO;
 import com.readjoy.readjoyapi.repository.BookRepository;
 import com.readjoy.readjoyapi.repository.CategoryRepository;
+import com.readjoy.readjoyapi.repository.ResourceRepository;
 import com.readjoy.readjoyapi.service.BookService;
-import jakarta.annotation.Resource;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 /**
  * @author 13296
@@ -32,15 +38,42 @@ import java.util.List;
 @Service
 public class BookServiceImpl extends ServiceImpl<BookMapper, Book> implements BookService {
 
-    @Resource
+    @jakarta.annotation.Resource
     private BookRepository bookRepository;
-    @Resource
+    @jakarta.annotation.Resource
     private CategoryRepository categoryRepository;
-    @Resource
+    @jakarta.annotation.Resource
+    private ResourceRepository resourceRepository;
+    @jakarta.annotation.Resource
     private LocalFileUtil localFileUtil;
 
     @Override
-    public IPage<BookVO> getPageByDTO(SelectBookDTO dto) {
+    public IPage<BookVO> getAdminPageByDTO(SelectBookDTO<BookVO> dto) {
+        return bookRepository.selectPageByDTO(dto);
+
+    }
+
+    @Override
+    public IPage<BookDetailVO> getBookDetailListByPage(SelectBookDTO<BookDetailVO> dto) {
+        final IPage<BookDetailVO> pageInfo = bookRepository.getAdminPageByDTO(dto);
+        // 单独查询资源列表
+        final List<Integer> collect = pageInfo.getRecords().stream().map(BookDetailVO::getBookId).toList();
+        if (!collect.isEmpty()) {
+            final Stream<AdminResourceVO> resourceList = resourceRepository.list(new LambdaQueryWrapper<Resource>()
+                    .in(Resource::getBookId, collect)).stream().map(AdminResourceVO::toVO);
+            // 将资源列表转换为以BookId为键的Map，以提高查找效率
+            Map<Integer, List<AdminResourceVO>> resourceMap = resourceList.collect(Collectors.groupingBy(AdminResourceVO::getBookId));
+            // 遍历pageInfo中的记录，并将对应的资源列表设置到BookDetailVO中
+            pageInfo.getRecords().forEach(bookDetailVO -> {
+                List<AdminResourceVO> resources = resourceMap.getOrDefault(bookDetailVO.getBookId(), Collections.emptyList());
+                bookDetailVO.setResourceList(resources);
+            });
+        }
+        return pageInfo;
+    }
+
+    @Override
+    public IPage<BookVO> getPageByDTO(SelectBookDTO<BookVO> dto) {
         return bookRepository.selectPageByDTO(dto);
     }
 
@@ -91,15 +124,13 @@ public class BookServiceImpl extends ServiceImpl<BookMapper, Book> implements Bo
         AssertUtil.isTrue(longs.length > 0, "删除图书不能为空！");
         AssertUtil.isTrue(longs.length < 500, "一次最多只能删除500本图书！");
         // 1. 查询图书封面
-        final List<Book> list = bookRepository.list(new LambdaQueryWrapper<Book>().in(Book::getBookId, Arrays.asList(longs))
-                .select(Book::getCoverImageUrl, Book::getBookId));
+        final List<Book> list = bookRepository.list(new LambdaQueryWrapper<Book>().in(Book::getBookId, Arrays.asList(longs)).select(Book::getCoverImageUrl, Book::getBookId));
 
         final Integer integer = bookRepository.batchDelete(longs);
         AssertUtil.isTrue(integer > 0 && integer == longs.length, "删除图书失败，请稍后重试！");
         // 2. 删除图书封面
         for (Book book : list)
-            if (StringUtils.isNotBlank(book.getCoverImageUrl()))
-                localFileUtil.deleteFile(book.getCoverImageUrl());
+            if (StringUtils.isNotBlank(book.getCoverImageUrl())) localFileUtil.deleteFile(book.getCoverImageUrl());
         return integer;
     }
